backend/campus_backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
DateTime==5.5
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.15.1
filelock==3.13.4
fsspec==2024.3.1
groq==0.5.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
openai==1.22.0
packaging==24.0
pinecone-client==3.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
setuptools==69.2.0
sniffio==1.3.1
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.43.0
whitenoise==6.6.0
zope.interface==6.2
